refactor: create utxo_pool.rs with related data structures
