+++ data_structures/src/chain.rs
@@ -25,7 +25,6 @@ use witnet_crypto::{
 };
 use witnet_protected::Protected;
 use witnet_reputation::{ActiveReputationSet, TotalReputationSet};
-use witnet_util::timestamp::get_timestamp;
 
 use crate::{
     chain::Signature::Secp256k1,
@@ -41,6 +40,7 @@ use crate::{
         CommitTransaction, DRTransaction, DRTransactionBody, MintTransaction, RevealTransaction,
         TallyTransaction, Transaction, TxInclusionProof, VTTransaction,
     },
+    utxo_pool::{OwnUnspentOutputsPool, UnspentOutputsPool},
     vrf::{BlockEligibilityClaim, DataRequestEligibilityClaim},
 };
 
@@ -2646,241 +2646,6 @@ pub enum DataRequestStage {
 impl Default for DataRequestStage {
     fn default() -> Self {
         DataRequestStage::COMMIT
-    }
-}
-
-/// Unspent Outputs Pool
-#[derive(Clone, Debug, Default, Eq, PartialEq, Serialize, Deserialize)]
-pub struct UnspentOutputsPool {
-    /// Map of output pointer to a tuple of:
-    /// * Value transfer output
-    /// * The number of the block that included the transaction
-    ///   (how many blocks were consolidated before this one).
-    map: HashMap<OutputPointer, (ValueTransferOutput, u32)>,
-}
-
-impl UnspentOutputsPool {
-    pub fn get<Q: ?Sized>(&self, k: &Q) -> Option<&ValueTransferOutput>
-    where
-        OutputPointer: std::borrow::Borrow<Q>,
-        Q: std::hash::Hash + Eq,
-    {
-        self.map.get(k).map(|(vt, _n)| vt)
-    }
-
-    pub fn contains_key<Q: ?Sized>(&self, k: &Q) -> bool
-    where
-        OutputPointer: std::borrow::Borrow<Q>,
-        Q: std::hash::Hash + Eq,
-    {
-        self.map.contains_key(k)
-    }
-
-    pub fn insert(
-        &mut self,
-        k: OutputPointer,
-        v: ValueTransferOutput,
-        block_number: u32,
-    ) -> Option<(ValueTransferOutput, u32)> {
-        self.map.insert(k, (v, block_number))
-    }
-
-    pub fn remove(&mut self, k: &OutputPointer) -> Option<(ValueTransferOutput, u32)> {
-        self.map.remove(k)
-    }
-
-    pub fn drain(
-        &mut self,
-    ) -> std::collections::hash_map::Drain<OutputPointer, (ValueTransferOutput, u32)> {
-        self.map.drain()
-    }
-
-    pub fn iter(
-        &self,
-    ) -> std::collections::hash_map::Iter<OutputPointer, (ValueTransferOutput, u32)> {
-        self.map.iter()
-    }
-
-    /// Returns the number of the block that included the transaction referenced
-    /// by this OutputPointer. The difference between that number and the
-    /// current number of consolidated blocks is the "collateral age".
-    pub fn included_in_block_number(&self, k: &OutputPointer) -> Option<Epoch> {
-        self.map.get(k).map(|(_vt, n)| *n)
-    }
-}
-
-/// List of unspent outputs that can be spent by this node
-#[derive(Clone, Debug, Default, Eq, PartialEq, Serialize, Deserialize)]
-pub struct OwnUnspentOutputsPool {
-    /// Map of output pointer to timestamp
-    /// Those UTXOs have a timestamp value to avoid double spending
-    map: HashMap<OutputPointer, u64>,
-}
-
-impl OwnUnspentOutputsPool {
-    pub fn new() -> Self {
-        Self {
-            map: HashMap::default(),
-        }
-    }
-    pub fn get<Q: ?Sized>(&self, k: &Q) -> Option<&u64>
-    where
-        OutputPointer: std::borrow::Borrow<Q>,
-        Q: std::hash::Hash + Eq,
-    {
-        self.map.get(k)
-    }
-
-    pub fn get_mut<Q: ?Sized>(&mut self, k: &Q) -> Option<&mut u64>
-    where
-        OutputPointer: std::borrow::Borrow<Q>,
-        Q: std::hash::Hash + Eq,
-    {
-        self.map.get_mut(k)
-    }
-
-    pub fn contains_key<Q: ?Sized>(&self, k: &Q) -> bool
-    where
-        OutputPointer: std::borrow::Borrow<Q>,
-        Q: std::hash::Hash + Eq,
-    {
-        self.map.contains_key(k)
-    }
-
-    pub fn insert(&mut self, k: OutputPointer, v: u64) -> Option<u64> {
-        self.map.insert(k, v)
-    }
-
-    pub fn remove(&mut self, k: &OutputPointer) -> Option<u64> {
-        self.map.remove(k)
-    }
-
-    pub fn drain(&mut self) -> std::collections::hash_map::Drain<OutputPointer, u64> {
-        self.map.drain()
-    }
-
-    pub fn iter(&self) -> std::collections::hash_map::Iter<OutputPointer, u64> {
-        self.map.iter()
-    }
-
-    pub fn keys(&self) -> std::collections::hash_map::Keys<OutputPointer, u64> {
-        self.map.keys()
-    }
-
-    pub fn len(&self) -> usize {
-        self.map.len()
-    }
-
-    pub fn is_empty(&self) -> bool {
-        self.map.is_empty()
-    }
-
-    /// Method to sort own_utxos by value
-    pub fn sort(&self, all_utxos: &UnspentOutputsPool, bigger_first: bool) -> Vec<OutputPointer> {
-        self.keys()
-            .sorted_by_key(|o| {
-                let value = all_utxos.get(o).map(|vt| i128::from(vt.value)).unwrap_or(0);
-
-                if bigger_first {
-                    -value
-                } else {
-                    value
-                }
-            })
-            .cloned()
-            .collect()
-    }
-}
-
-/// Strategy to sort our own unspent outputs pool
-#[derive(Clone, Debug, Serialize, Deserialize, Eq, PartialEq, Hash)]
-pub enum UtxoSelectionStrategy {
-    Random,
-    BigFirst,
-    SmallFirst,
-}
-
-impl Default for UtxoSelectionStrategy {
-    fn default() -> Self {
-        UtxoSelectionStrategy::Random
-    }
-}
-
-#[derive(Clone, Debug, Serialize, Deserialize)]
-pub struct UtxoMetadata {
-    pub output_pointer: OutputPointer,
-    pub value: u64,
-    pub timelock: u64,
-    pub utxo_mature: bool,
-}
-
-/// Information about our own UTXOs
-#[derive(Clone, Debug, Serialize, Deserialize)]
-pub struct UtxoInfo {
-    /// Vector of OutputPointer with their values, time_locks and if it is ready for collateral
-    pub utxos: Vec<UtxoMetadata>,
-    /// Minimum collateral from consensus constants
-    pub collateral_min: u64,
-}
-
-#[allow(clippy::cast_sign_loss)]
-fn create_utxo_metadata(
-    vto: &ValueTransferOutput,
-    o: &OutputPointer,
-    all_utxos: &UnspentOutputsPool,
-    block_number_limit: u32,
-) -> UtxoMetadata {
-    let now = get_timestamp() as u64;
-    let timelock = if vto.time_lock >= now {
-        vto.time_lock
-    } else {
-        0
-    };
-    let utxo_mature: bool = all_utxos.included_in_block_number(o).unwrap() <= block_number_limit;
-
-    UtxoMetadata {
-        output_pointer: o.clone(),
-        value: vto.value,
-        timelock,
-        utxo_mature,
-    }
-}
-
-/// Get Utxo Information
-#[allow(clippy::cast_sign_loss)]
-pub fn get_utxo_info(
-    pkh: Option<PublicKeyHash>,
-    own_utxos: &OwnUnspentOutputsPool,
-    all_utxos: &UnspentOutputsPool,
-    collateral_min: u64,
-    block_number_limit: u32,
-) -> UtxoInfo {
-    let utxos = if let Some(pkh) = pkh {
-        all_utxos
-            .iter()
-            .filter_map(|(o, (vto, _))| {
-                if vto.pkh == pkh {
-                    Some(create_utxo_metadata(vto, o, all_utxos, block_number_limit))
-                } else {
-                    None
-                }
-            })
-            .collect()
-    } else {
-        // Read your own UtxoInfo is cheaper than from other pkhs
-        own_utxos
-            .iter()
-            .filter_map(|(o, _)| {
-                all_utxos
-                    .get(o)
-                    .map(|vto| create_utxo_metadata(vto, o, all_utxos, block_number_limit))
-            })
-            .collect()
-    };
-
-    UtxoInfo {
-        utxos,
-        collateral_min,
     }
 }
